# Add optimized path for power-of-two-sized buffers