--- a/repo.yml
+++ b/repo.yml
@@ -1,0 +1,1 @@
+type: 'generic'