@lmbarros 2021-05-17
FWIW, I also opened a PR to try to get all our optimizations upstream: https://github.com/armon/circbuf/pull/5